requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
certifi==2022.6.15
charset-normalizer==2.1.0
cos-python-sdk-v5==1.9.20
crcmod==1.7
dicttoxml==1.7.4
Django==4.0.6
django-cors-headers==3.13.0
django-crontab==0.7.1
django-model-utils==4.3.1
django-notifications-hq
django-rest-framework==0.1.0
djangorestframework==3.13.1
idna==3.3
jsonfield==3.1.0
mysql==0.0.3
mysqlclient==2.1.1
Pillow==9.2.0
pycryptodome==3.15.0
PyMySQL==1.0.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
urllib3==1.26.11

notifications~=0.3.2
joblib~=1.2.0
pandas~=2.0.0
jieba~=0.42.1
scikit-learn~=1.2.2
channels~=4.0.0